+++ package.json
@@ -9,7 +9,7 @@
     "watch": "node --watch app.js",
     "test": "jest --watchAll --runInBand --verbose",
     "ci_tests": "jest --runInBand --verbose",
-    "seed": "node seedDB.js"
+    "seed": "node ./seed/seedDB.js"
   },
   "repository": {
     "type": "git",
