+++ seed/seedDB.js
@@ -1,0 +1,4 @@
+import { initDB } from "./seed.js";
+import { closeConnection } from "../config/mongoConnection.js";
+await initDB();
+await closeConnection();
