+++ seed/users.json
@@ -1,0 +1,21 @@
+[{
+  "_id": {
+    "$oid": "507f191e810c19729de860ea"
+  },
+  "firstName": "Siva Anand",
+  "lastName": "Sivakumar",
+  "username": "sivaanand",
+  "password": "$2b$04$I./5JP7RxQcjqdR4BpiiO.KqChkRwZSVVveTN2eFhj3ONuBelo7Me",
+  "bio": "Hello! I am Siva Anand Sivakumar, an admin for the photography spot finder!",
+  "email": "ssivakum1@stevens.edu",
+  "favoriteSpots": [],
+  "comments": [],
+  "ratings": [],
+  "spots": [],
+  "contestSubmissions": [],
+  "contestVote": [],
+  "sportReports": [],
+  "commentReports": [],
+  "contestReports": [],
+  "role": "admin"
+}]
