--- seedDB.js
@@ -1,4 +1,0 @@
-import { initDB } from "./seed.js";
-import { closeConnection } from "./config/mongoConnection.js";
-await initDB();
-await closeConnection();
